feat(jans-auth-server): introduce acr mappings and prefixed acr - "agama_<flow name> " #8216
https://github.com/JanssenProject/jans/issues/8216
