+++ jans-auth-server/server/src/main/java/io/jans/as/server/authorize/ws/rs/AuthorizeRestWebServiceValidator.java
@@ -9,33 +9,23 @@ package io.jans.as.server.authorize.ws.rs;
 import com.google.common.base.Strings;
 import io.jans.as.common.model.registration.Client;
 import io.jans.as.common.model.session.SessionId;
-import io.jans.as.common.model.session.SessionIdState;
 import io.jans.as.common.util.RedirectUri;
 import io.jans.as.model.authorize.AuthorizeErrorResponseType;
 import io.jans.as.model.authzdetails.AuthzDetails;
-import io.jans.as.model.common.Prompt;
 import io.jans.as.model.common.ResponseMode;
 import io.jans.as.model.common.ResponseType;
 import io.jans.as.model.configuration.AppConfiguration;
 import io.jans.as.model.crypto.signature.SignatureAlgorithm;
 import io.jans.as.model.error.ErrorResponseFactory;
 import io.jans.as.model.exception.InvalidJwtException;
-import io.jans.as.model.util.Util;
 import io.jans.as.server.model.authorize.AuthorizeParamsValidator;
 import io.jans.as.server.model.authorize.JwtAuthorizationRequest;
 import io.jans.as.server.model.common.DeviceAuthorizationCacheControl;
-import io.jans.as.server.model.exception.AcrChangedException;
 import io.jans.as.server.model.exception.InvalidRedirectUrlException;
-import io.jans.as.server.security.Identity;
 import io.jans.as.server.service.*;
-import io.jans.as.server.service.external.ExternalAuthenticationService;
 import io.jans.as.server.service.external.ExternalAuthzDetailTypeService;
-import io.jans.as.server.service.external.session.SessionEvent;
-import io.jans.as.server.service.external.session.SessionEventType;
 import io.jans.as.server.util.RedirectUtil;
 import io.jans.as.server.util.ServerUtil;
-import io.jans.model.AuthenticationScriptUsageType;
-import io.jans.model.custom.script.conf.CustomScriptConfiguration;
 import io.jans.orm.exception.EntryPersistenceException;
 import jakarta.ejb.Stateless;
 import jakarta.inject.Inject;
@@ -84,16 +74,7 @@ public class AuthorizeRestWebServiceValidator {
     private AppConfiguration appConfiguration;
 
     @Inject
-    private SessionIdService sessionIdService;
-
-    @Inject
-    private Identity identity;
-
-    @Inject
     private ExternalAuthzDetailTypeService externalAuthzDetailTypeService;
-
-    @Inject
-    private ExternalAuthenticationService externalAuthenticationService;
 
     public Client validateClient(String clientId, String state) {
         return validateClient(clientId, state, false);
@@ -403,60 +384,6 @@ public class AuthorizeRestWebServiceValidator {
         if (isTrue(appConfiguration.getRequirePkce()) && Strings.isNullOrEmpty(codeChallenge)) {
             log.error("PKCE is required but code_challenge is blank.");
             throw redirectUriResponse.createWebException(AuthorizeErrorResponseType.INVALID_REQUEST);
-        }
-    }
-
-    public void validateAcrs(AuthzRequest authzRequest, Client client) throws AcrChangedException {
-        if (!client.getAttributes().getAuthorizedAcrValues().isEmpty() &&
-                !client.getAttributes().getAuthorizedAcrValues().containsAll(authzRequest.getAcrValuesList())) {
-            throw authzRequest.getRedirectUriResponse().createWebException(AuthorizeErrorResponseType.INVALID_REQUEST,
-                    "Restricted acr value request, please review the list of authorized acr values for this client");
-        }
-
-        checkAcrScriptIsAvailable(authzRequest);
-        checkAcrChanged(authzRequest, identity.getSessionId()); // check after redirect uri is validated
-    }
-
-    public void checkAcrScriptIsAvailable(AuthzRequest authzRequest) {
-        final String acrValues = authzRequest.getAcrValues();
-        if (StringUtils.isBlank(acrValues)) {
-            return; // nothing to validate
-        }
-
-        if (Util.isBuiltInPasswordAuthn(acrValues)) {
-            return; // no need for script for built-in "simple_password_auth"
-        }
-
-        CustomScriptConfiguration script = externalAuthenticationService.determineCustomScriptConfiguration(AuthenticationScriptUsageType.INTERACTIVE, authzRequest.getAcrValuesList());
-        if (script == null) {
-            String msg = String.format("Unable to find script for acr: %s. Send error: %s",
-                    acrValues, AuthorizeErrorResponseType.UNMET_AUTHENTICATION_REQUIREMENTS.getParameter());
-            log.debug(msg);
-            throw authzRequest.getRedirectUriResponse().createWebException(AuthorizeErrorResponseType.UNMET_AUTHENTICATION_REQUIREMENTS, msg);
-        }
-    }
-
-    private void checkAcrChanged(AuthzRequest authzRequest, SessionId sessionUser) throws AcrChangedException {
-        try {
-            sessionIdService.assertAuthenticatedSessionCorrespondsToNewRequest(sessionUser, authzRequest.getAcrValues());
-        } catch (AcrChangedException e) { // Acr changed
-            //See https://github.com/GluuFederation/oxTrust/issues/797
-            if (e.isForceReAuthentication()) {
-                if (!authzRequest.getPromptList().contains(Prompt.LOGIN)) {
-                    log.info("ACR is changed, adding prompt=login to prompts");
-                    authzRequest.addPrompt(Prompt.LOGIN);
-
-                    sessionUser.setState(SessionIdState.UNAUTHENTICATED);
-                    sessionUser.getSessionAttributes().put("prompt", authzRequest.getPrompt());
-                    if (!sessionIdService.persistSessionId(sessionUser)) {
-                        log.trace("Unable persist session_id, trying to update it.");
-                        sessionIdService.updateSessionId(sessionUser);
-                    }
-                    sessionIdService.externalEvent(new SessionEvent(SessionEventType.UNAUTHENTICATED, sessionUser));
-                }
-            } else {
-                throw e;
-            }
         }
     }
