+++ jans-auth-server/server/src/test/java/io/jans/as/server/authorize/ws/rs/AuthorizeRestWebServiceValidatorTest.java
@@ -5,14 +5,11 @@ import io.jans.as.common.model.session.SessionId;
 import io.jans.as.common.util.RedirectUri;
 import io.jans.as.model.configuration.AppConfiguration;
 import io.jans.as.model.error.ErrorResponseFactory;
-import io.jans.as.server.security.Identity;
-import io.jans.as.server.service.*;
-import io.jans.as.server.service.external.ExternalAuthenticationService;
+import io.jans.as.server.service.ClientService;
+import io.jans.as.server.service.DeviceAuthorizationService;
+import io.jans.as.server.service.RedirectUriResponse;
+import io.jans.as.server.service.RedirectionUriService;
 import io.jans.as.server.service.external.ExternalAuthzDetailTypeService;
-import io.jans.model.AuthenticationScriptUsageType;
-import io.jans.model.custom.script.conf.CustomScriptConfiguration;
-import io.jans.model.custom.script.model.CustomScript;
-import io.jans.model.custom.script.type.auth.DummyPersonAuthenticationType;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.ws.rs.WebApplicationException;
 import org.mockito.InjectMocks;
@@ -23,7 +20,6 @@ import org.testng.annotations.Listeners;
 import org.testng.annotations.Test;
 
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.HashSet;
 
 import static org.mockito.Mockito.*;
@@ -57,60 +53,7 @@ public class AuthorizeRestWebServiceValidatorTest {
     private AppConfiguration appConfiguration;
 
     @Mock
-    private SessionIdService sessionIdService;
-
-    @Mock
-    private Identity identity;
-
-    @Mock
     private ExternalAuthzDetailTypeService externalAuthzDetailTypeService;
-
-    @Mock
-    private ExternalAuthenticationService externalAuthenticationService;
-
-    @Test
-    public void checkAcrScriptIsAvailable_forBlankAcr_shouldPass() {
-        AuthzRequest authzRequest = new AuthzRequest();
-
-        authorizeRestWebServiceValidator.checkAcrScriptIsAvailable(authzRequest);
-    }
-
-    @Test
-    public void checkAcrScriptIsAvailable_forBuildInAcr_shouldPass() {
-        AuthzRequest authzRequest = new AuthzRequest();
-        authzRequest.setAcrValues("simple_password_auth");
-
-        authorizeRestWebServiceValidator.checkAcrScriptIsAvailable(authzRequest);
-    }
-
-    @Test
-    public void checkAcrScriptIsAvailable_whenScriptIsAvailable_shouldPass() {
-        AuthzRequest authzRequest = new AuthzRequest();
-        authzRequest.setAcrValues("my_acr");
-
-        final CustomScriptConfiguration script = new CustomScriptConfiguration(new CustomScript(), new DummyPersonAuthenticationType(), new HashMap<>());
-        when(externalAuthenticationService.determineCustomScriptConfiguration(AuthenticationScriptUsageType.INTERACTIVE, authzRequest.getAcrValuesList())).thenReturn(script);
-
-        authorizeRestWebServiceValidator.checkAcrScriptIsAvailable(authzRequest);
-    }
-
-    @Test
-    public void checkAcrScriptIsAvailable_whenScriptIsNotAvailable_shouldFail() {
-        RedirectUri redirectUri = mock(RedirectUri.class);
-        when(redirectUri.toString()).thenReturn("http://rp.com");
-
-        AuthzRequest authzRequest = new AuthzRequest();
-        authzRequest.setAcrValues("my_acr");
-        authzRequest.setRedirectUriResponse(new RedirectUriResponse(redirectUri, "", mock(HttpServletRequest.class), mock(ErrorResponseFactory.class)));
-
-        try {
-            authorizeRestWebServiceValidator.checkAcrScriptIsAvailable(authzRequest);
-        } catch (WebApplicationException e) {
-            return;
-        }
-
-        fail("Script is not available but exception is not thrown.");
-    }
 
     @Test
     public void validateRequestParameterSupported_whenRequestIsEmpty_shouldPass() {
